refactor:  점주, 손님 테이블 통합으로 회원 정보 조회 기능 수정 - UserEditParam -> UserEditDto로 변경 - UserParam -> UserInfoDto로 변경 - 서비스에서 getUserInfo에서 Merchant 관련 루틴 삭제
