+++ src/main/java/fingerorder/webapp/dto/UserInfoDto.java
@@ -1,11 +1,11 @@
-package fingerorder.webapp.parameter;
+package fingerorder.webapp.dto;
 
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class UserParam {
+public class UserInfoDto {
 	private String email;
 	private String type;
 }
